+++ userscript/src/settings/settingsTab.ts
@@ -1,7 +1,8 @@
 import config from "../config";
 import { saveSettings, SETTINGS } from "./settings";
+import hookGraphicsSettingsTab from "./settingsTabGraphics";
+import hookSoundSettingsTab from "./settingsTabSound";
 import getTexTab from "./settingsTabTex";
-import getUtilTab from "./settingsTabUtil";
 
 export default function settingsTab() {
   // retry panel inject if failed
@@ -56,11 +57,9 @@ export default function settingsTab() {
       });
       pan.addChild(pan[`${name}ButtonBackground`]);
     }
-    newTab("NinjaIOUtils", "util", 302, getUtilTab());
-    newTab("Texture Pack", "tex", 418, getTexTab());
+    newTab("Texture Pack", "tex", 302, getTexTab());
     return pan;
   }
-  SettingsPanel.Tabs.UTIL = "util";
   SettingsPanel.Tabs.TEX = "tex";
   const oldX = app.menu.settingsPanel.x,
     oldY = app.menu.settingsPanel.y;
@@ -99,53 +98,6 @@ export default function settingsTab() {
     tab._events.mousedown.shift();
   });
 
-  app.menu.settingsPanel.graphicsTab.addChild(
-    (app.menu.settingsPanel.graphicsTab.uiScaler = new Slider(
-      "chatOpacity",
-      "UI scale",
-      SETTINGS.uiScale || 0.4,
-      2,
-      0.4
-    ))
-  );
-  if (!SETTINGS.uiScale) app.menu.settingsPanel.graphicsTab.uiScaler.valueLabel.text = "default";
-  app.menu.settingsPanel.graphicsTab.uiScaler.x = app.menu.settingsPanel.graphicsTab.enableAA.x;
-  app.menu.settingsPanel.graphicsTab.uiScaler.y =
-    app.menu.settingsPanel.graphicsTab.enableAA.y +
-    app.menu.settingsPanel.graphicsTab.enableAA.height +
-    14;
-  app.menu.settingsPanel.graphicsTab.uiScaler.on(Slider.CHANGE, (b: number) => {
-    b = Math.round(b * 10) / 10;
-    if (b == 0.4) {
-      app.menu.settingsPanel.graphicsTab.uiScaler.valueLabel.text = "default";
-      b = 0;
-    }
-    if (b == SETTINGS.uiScale) return;
-    App.NUIScale = SETTINGS.uiScale = b;
-    saveSettings();
-    app.onResize();
-
-    const conf = document.createElement("div");
-    const confy = document.createElement("button");
-    confy.innerHTML = "OK";
-    confy.onclick = () => conf.remove();
-    conf.appendChild(confy);
-    const confn = document.createElement("button");
-    confn.innerHTML = "Undo";
-    confn.onclick = () => {
-      App.NUIScale = SETTINGS.uiScale = 0;
-      saveSettings();
-      app.onResize();
-      conf.remove();
-    };
-    confn.style.marginLeft = "0.3rem";
-    conf.appendChild(confn);
-    conf.style.position = "absolute";
-    conf.style.top = "50%";
-    conf.style.left = "50%";
-    conf.style.transform = "translate(-50%,-50%)";
-    conf.style.backgroundColor = "rgba(0,0,0,0.8)";
-    conf.style.padding = "0.5rem";
-    document.body.appendChild(conf);
-  });
+  hookSoundSettingsTab();
+  hookGraphicsSettingsTab();
 }